Make inttime_t 64 bit

+++ libgadget/types.h
@@ -6,7 +6,7 @@
 /*Define some useful types*/
 
 typedef uint32_t binmask_t;
-typedef int32_t inttime_t;
+typedef int64_t inttime_t;
 
 #define BINMASK_ALL ((uint32_t) (-1))
 #define BINMASK(i) (1u << i)
